requirements.txt
blinker==1.7.0
certifi==2024.2.2
charset-normalizer==3.3.2
click==8.1.7
contourpy==1.2.0
cycler==0.12.1
Flask==3.0.2
fonttools==4.49.0
idna==3.6
importlib-metadata==7.0.1
importlib_resources==6.1.2
itsdangerous==2.1.2
Jinja2==3.1.3
kiwisolver==1.4.5
MarkupSafe==2.1.5
matplotlib==3.8.3
numpy==1.26.4
packaging==23.2
pillow==10.2.0
pyparsing==3.1.1
python-dateutil==2.8.2
requests==2.31.0
six==1.16.0
urllib3==2.2.1
Werkzeug==3.0.1
zipp==3.17.0
